requirements.txt
dash==3.0.1
dash-bootstrap-components==2.0.0
et_xmlfile==2.0.0
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
python-dotenv==1.1.0
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
plotly==6.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
SQLAlchemy==2.0.40